Moved cucumber BDD features to features package
